requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
kombu==5.4.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
